Methods

Functions

Algorithmic Methods

These are the high-level algorithmic methods:

POD.presolve — Function.

presolve(m::PODNonlinearModel)

Algorithmic function that ask a loaded POD problem to perform the presolving process. It first conduct a local solve for a initial starting point followed by the bound tightening process. In cases when a local solve is infeasible, the bound tightening will continue with the bound tightening. However, as the bound tightening finishes, a second try of the local solve will be performed on a tigher domain for a initial feasible solution. The problem will be position the partitioning to the initial feasible solution or a relaxation root McCormick solution.

See Algorithm ... for details implementation.

source

POD.global_solve — Function.

global_solve(m::PODNonlinearModel)

Perform the global algorithm that is based on the adaptive conexification scheme. This iterative algorithm loops over bounding_solve and local_solve for converging lower bound (relaxed problem) and upper bound (feasible problem). Each bounding_solve provides a lower bound solution that is used as a partioning point for next iteration (this feature can be modified given different add_discretization). Each local_solve provides a local serach of incumbent feasible solution. The algrithm terminates given time limits, optimality condition, or iteration limits.

The algorithm is can be reformed when add_discretization is replaced with user-defined functional input. For example, this algorithm can easily be reformed as a uniform-partitioning algorithm in other literature.

source

POD.local_solve — Function.

local_solve(m::PODNonlinearModel, presolve::Bool=false)

Perform a local NLP or MINLP solve to detect feasible solution. When presolve=true, this function is utilized in the presolve for a MINLP local solve. Otherwise, a NLP problem is solved by fixing the integer variables base on the lower bound solution from bounding_solve.

source

POD.bounding_solve — Function.

bounding_solve(m::PODNonlinearModel; kwargs...)

This is a solving process usually deal with a MIP or MIQCP problem for lower bounds of problems. It solves the problem built upon a convexification base on a discretization Dictionary of some variables. The convexification utilized is Tighten McCormick scheme. See create_bounding_mip for more details of the problem solved here.

source

Utility Methods