Functions

High-level Algorithmic Operations

These are the high-level algorithmic methods:

POD.presolve — Function.

presolve(m::PODNonlinearModel)

Function that perfoms a presolve on the user-supplied nonlinear program. The presolve first provides the model to a local solver to obtain an initial feasible solution. If the local solver returns a feasible solution then the objective value of the feasible solution is used in conjunction with the original model for a bound-tightening procedure. If the local solver reports infeasibililty the bound-tightening procedure does not use the objective value. Furthermore, in this case, a second local solve is attempted using the tightened bounds.

If a local solution is not obtained eved after the second solve then an initial McCormick solve is performed. The local solution (if available) or the initial McCormick solution (if infeasible after two local solve tries) is then used to partition the variables for the subsequent Adaptive Multivariate Partitioning algorithm iterations.

source

POD.global_solve — Function.

global_solve(m::PODNonlinearModel)

Perform the global algorithm that is based on the adaptive conexification scheme. This iterative algorithm loops over bounding_solve and local_solve for converging lower bound (relaxed problem) and upper bound (feasible problem). Each bounding_solve provides a lower bound solution that is used as a partioning point for next iteration (this feature can be modified given different add_discretization). Each local_solve provides a local serach of incumbent feasible solution. The algrithm terminates given time limits, optimality condition, or iteration limits.

The algorithm is can be reformed when add_discretization is replaced with user-defined functional input. For example, this algorithm can easily be reformed as a uniform-partitioning algorithm in other literature.

source

POD.local_solve — Function.

local_solve(m::PODNonlinearModel, presolve::Bool=false)

Perform a local NLP or MINLP solve to obtain a feasible solution. The presolve option is set to true when the function is invoked in presolve. Otherwise, the function is invoked from bounding_solve.

source

POD.bounding_solve — Function.

bounding_solve(m::PODNonlinearModel; kwargs...)

This is a solving process usually deal with a MIP or MIQCP problem for lower bounds of problems. It solves the problem built upon a convexification base on a discretization Dictionary of some variables. The convexification utilized is Tighten McCormick scheme. See create_bounding_mip for more details of the problem solved here.

source

Adapative Partitioning Methods

POD.create_bounding_mip — Function.

create_bounding_mip(m::PODNonlinearModel; use_discretization::Dict)

Set up a JuMP MILP bounding model base on variable domain partitioning information stored in use_discretization. By default, if use_discretization is not provided, it will use m.discretizations store in the POD model. The basic idea of this MILP bounding model is to use Tighten McCormick to convexify the original Non-convex region. Among all presented partitionings, the bounding model will choose one specific partition as the lower bound solution. The more partitions there are, the better or finer bounding model relax the original MINLP while the more efforts required to solve this MILP is required.

This function is implemented in the following manner:

* [`post_amp_vars`](@ref): post original and lifted variables
* [`post_amp_lifted_constraints`](@ref): post original and lifted constraints
* [`post_amp_lifted_obj`](@ref): post original or lifted objective function
* [`post_amp_mccormick`](@ref): post Tighen McCormick variables and constraints base on `discretization` information

More specifically, the Tightening McCormick used here can be genealized in the following mathematcial formulation. Consider a nonlinear term

\[\begin{subequations} \begin{align} &\widehat{x_{ij}} \geq (\mathbf{x}_i^l\cdot\hat{\mathbf{y}}_i) x_j + (\mathbf{x}_j^l\cdot\hat{\mathbf{y}}_j) x_i - (\mathbf{x}_i^l\cdot\hat{\mathbf{y}}_i)(\mathbf{x}_j^l\cdot\hat{\mathbf{y}}_j) \\ &\widehat{x_{ij}} \geq (\mathbf{x}_i^u\cdot\hat{\mathbf{y}}_i) x_j + (\mathbf{x}_j^u\cdot\hat{\mathbf{y}}_j) x_i - (\mathbf{x}_i^u\cdot\hat{\mathbf{y}}_i)(\mathbf{x}_j^u\cdot\hat{\mathbf{y}}_j) \\ &\widehat{x_{ij}} \leq (\mathbf{x}_i^l\cdot\hat{\mathbf{y}}_i) x_j + (\mathbf{x}_j^u\cdot\hat{\mathbf{y}}_j) x_i - (\mathbf{x}_i^l\cdot\hat{\mathbf{y}}_i)(\mathbf{x}_j^u\cdot\hat{\mathbf{y}}_j) \\ &\widehat{x_{ij}} \leq (\mathbf{x}_i^u\cdot\hat{\mathbf{y}}_i) x_j + (\mathbf{x}_j^l\cdot\hat{\mathbf{y}}_j) x_i - (\mathbf{x}_i^u\cdot\hat{\mathbf{y}}_i)(\mathbf{x}_j^l\cdot\hat{\mathbf{y}}_j) \\ & \mathbf{x}_i^u\cdot\hat{\mathbf{y}}_i) \geq x_{i} \geq \mathbf{x}_i^l\cdot\hat{\mathbf{y}}_i) \\ & \mathbf{x}_j^u\cdot\hat{\mathbf{y}}_j) \geq x_{j} \geq \mathbf{x}_j^l\cdot\hat{\mathbf{y}}_j) \\ &\sum \hat{\mathbf{y}_i} = 1, \ \ \sum \hat{\mathbf{y}_j}_k = 1 \\ &\hat{\mathbf{y}}_i \in \{0,1\}, \hat{\mathbf{y}}_j \in \{0,1\} \end{align} \end{subequations}\]

source

POD.pick_vars_discretization — Function.

pick_vars_discretization(m::PODNonlinearModel)

This function helps pick the variables for discretization. The method chosen depends on user-inputs. In case when indices::Int is provided, the method is chosen as built-in method. Currently, there exist two built-in method:

* `max-cover(m.discretization_var_pick_algo=0, default)`: pick all variables involved in the non-linear term for discretization
* `min-vertex-cover(m.discretization_var_pick_algo=1)`: pick a minimum vertex cover for variables involved in non-linear terms so that each non-linear term is at least convexified

For advance usage, m.discretization_var_pick_algo allows ::Function inputs. User is required to perform flexible methods in choosing the non-linear variable. For more information, read more details at Hacking Solver.

source

POD.fix_domains — Function.

fix_domains(m::PODNonlinearModel)

This function is used to fix variables to certain domains during the local solve process in the global_solve. More specifically, it is used in local_solve to fix binary and integer variables to lower bound solutions and discretizing varibles to the active domain according to lower bound solution.

source

POD.min_vertex_cover — Function.

min_vertex_cover(m:PODNonlinearModel)

A built-in method for selecting variables for discretization.

source

POD.max_cover — Function.

max_cover(m:PODNonlinearModel)

A built-in method for selecting variables for discretization. It selects all variables in the nonlinear terms.

source

Presolve Methods

POD.bound_tightening — Function.

bound_tightening(m::PODNonlinearModel)

Entry point for the bound-tightening algorithm. The aim of the bound-tightening algorithm is to tighten the variable bounds, if possible.

Currently, two bounding tightening method is implemented minmax_bound_tightening.

* Bound-tightening with basic McCormick
* Bound-tightening with McCormick partitions: (3 partitions around the local feasible solution)
If no local feasible solution is obtained, the algorithm defaults to bound-tightening with basic McCormick

source

POD.minmax_bound_tightening — Function.

minmax_bound_tightening(m:PODNonlinearModel; use_bound::Bool=true, use_tmc::Bool)

This function implements the bound-tightening algorithm to tighten the variable bounds. It utilizes either the basic McCormick relaxation or the Tightened McCormick relaxation (TMC) to tighten the bounds. The TMC has additional binary variables for partitioning.

The algorithm as two main parameters. The first is the use_tmc, which when set to true invokes the algorithm on the TMC relaxation. The second parameter use_bound takes in the objective value of the local solve solution stored in best_sol. The use_bound option is set to true when the local solve is successful is obtaining a feasible solution and the bound-tightening is to be performed using the objective value of the feasible solution, else this parameter is set to false

Several other parameters are available for the presolve algorithm tuning. For more details, see Parameters.

source

POD.create_bound_tightening_model — Function.

create_bound_tightening_model(m::PODNonlinearModel, discretization::Dict, bound::Float64)

This function takes in the initial discretization information and builds a bound-tightening model. It is an algorithm specific function called by minmax_bound_tightening

source

POD.solve_bound_tightening_model — Function.

solve_bound_tightening_model(m::PODNonlinearModel)

A function that solves the min and max bound-tightening model.

source

POD.resolve_lifted_var_bounds — Function.

resolve_lifted_var_bounds(nonlinear_info::Dict, discretization::Dict)

For discretization to be performed, we do not allow for a variable being discretized to have infinite bounds. The lifted variables will have infinite bounds and the function infers bounds on these variables. This process can help speed up the subsequent solve in subsequent iterations.

source

Utility Methods

POD.update_var_bounds — Function.

update_var_bounds(m::PODNonlinearModel, discretization::Dict; len::Float64=length(keys(discretization)))

This function take in a dictionary-based discretization information and convert them into two bounds vectors (l_var, u_var) by picking the smallest and largest numbers. User can specify a certain length that may contains variables that is out of the scope of discretization.

Output::

l_var::Vector{Float64}, u_var::Vector{Float64}

source

POD.discretization_to_bounds — Function.

discretization_to_bounds(d::Dict, l::Int)

Same as update_var_bounds

source

POD.initialize_discretization — Function.

initialize_discretization(m::PODNonlinearModel)

This function initialize the dynamic discretization used for any bounding models. By default, it takes (.l_var_orig, .u_var_orig) as the base information. User is allowed to use alternative bounds for initializing the discretization dictionary. The output is a dictionary with MathProgBase variable indices keys attached to the :PODNonlinearModel.discretization.

source

POD.to_discretization — Function.

to_discretization(m::PODNonlinearModel, lbs::Vector{Float64}, ubs::Vector{Float64})

Utility functions to convert bounds vectors to Dictionary based structures that is more suitable for partition operations.

source

POD.flatten_discretization — Function.

flatten_discretization(discretization::Dict)

Utility functions to eliminate all partition on discretizing variable and keep the loose bounds.

source

POD.add_discretization — Function.

add_discretization(m::PODNonlinearModel; use_discretization::Dict, use_solution::Vector)

Basic built-in method used to add a new partition on feasible domains of discretizing variables. This method make modification in .discretization

Consider original partition [0, 3, 7, 9], where LB/any solution is 4. Use ^ as the new partition, "|" as the original partition

A case when discretize ratio = 4 | –––– | - ^ – * – ^ –– | –––– | 0 3 3.5 4 4.5 7 9

A special case when discretize ratio = 2 | –––– | –– * –– ^ –– | –––– | 0 3 4 5 7 9

There are two options for this function,

* `use_discretization(default=m.discretization)`:: to regulate which is the base to add new partitions on
* `use_solution(default=m.best_bound_sol)`:: to regulate which solution to use when adding new partitions on

This function belongs to the hackable group, which means it can be replaced by the user to change the behvaior of the solver.

source

POD.update_mip_time_limit — Function.

update_mip_time_limit(m::PODNonlinearModel)

An utility function used to dynamically regulate MILP solver time limits to fit POD solver time limits.

source

POD.fetch_timeleft_symbol — Function.

fetch_timeleft_symbol(m::PODNonlinearModel)

An utility function used to recognize differnt sub-solvers return the timelimit setup keywords.

source