Methods

Functions

High-level Algorithmic Operations

These are the high-level algorithmic methods:

Alpine.presolve — Function.

presolve(m::AlpineNonlinearModel)

source

Alpine.global_solve — Function.

global_solve(m::AlpineNonlinearModel)

Perform global optimization algorithm that is based on the adaptive piecewise convexification. This iterative algorithm loops over bounding_solve and local_solve until the optimality gap between the lower bound (relaxed problem with min. objective) and the upper bound (feasible problem) is within the user prescribed limits. Each bounding_solve provides a lower bound that serves as the partioning point for the next iteration (this feature can be modified given a different add_adaptive_partition). Each local_solve provides an incumbent feasible solution. The algorithm terminates when atleast one of these conditions are satisfied: time limit, optimality condition, or iteration limit.

source

Alpine.local_solve — Function.

local_solve(m::AlpineNonlinearModel, presolve::Bool=false)

Perform a local NLP or MINLP solve to obtain a feasible solution. The presolve option is set to true when the function is invoked in presolve. Otherwise, the function is invoked from bounding_solve.

source

Alpine.bounding_solve — Function.

bounding_solve(m::AlpineNonlinearModel; kwargs...)

This is a solving process usually deal with a MIP or MIQCP problem for lower bounds of problems. It solves the problem built upon a convexification base on a discretization Dictionary of some variables. The convexification utilized is Tighten McCormick scheme. See create_bounding_mip for more details of the problem solved here.

source

Adapative Partitioning Methods

create_bounding_mip
pick_disc_vars
fix_domains
min_vertex_cover
max_cover

Presolve Methods

Alpine.bound_tightening — Function.

bound_tightening(m::AlpineNonlinearModel)

Entry point for the bound-tightening algorithm. The aim of the bound-tightening algorithm is to tighten the variable bounds, if possible.

Currently, two bounding tightening method is implemented minmax_bound_tightening.

* Bound-tightening with basic McCormick
* Bound-tightening with McCormick partitions: (3 partitions around the local feasible solution)
If no local feasible solution is obtained, the algorithm defaults to bound-tightening with basic McCormick

source

Alpine.minmax_bound_tightening — Function.

minmax_bound_tightening(m:AlpineNonlinearModel; use_bound::Bool=true, use_tmc::Bool)

This function implements the bound-tightening algorithm to tighten the variable bounds. It utilizes either the basic McCormick relaxation or the Tightened McCormick relaxation (TMC) to tighten the bounds. The TMC has additional binary variables for partitioning.

The algorithm as two main parameters. The first is the use_tmc, which when set to true invokes the algorithm on the TMC relaxation. The second parameter use_bound takes in the objective value of the local solve solution stored in best_sol. The use_bound option is set to true when the local solve is successful is obtaining a feasible solution and the bound-tightening is to be performed using the objective value of the feasible solution, else this parameter is set to false

Several other parameters are available for the presolve algorithm tuning. For more details, see Parameters.

source

Alpine.create_bound_tightening_model — Function.

create_bound_tightening_model(m::AlpineNonlinearModel, discretization::Dict, bound::Float64)

This function takes in the initial discretization information and builds a bound-tightening model. It is an algorithm specific function called by minmax_bound_tightening

source

Alpine.solve_bound_tightening_model — Function.

solve_bound_tightening_model(m::AlpineNonlinearModel)

A function that solves the min and max bound-tightening model.

source

Alpine.resolve_var_bounds — Function.

resolve_var_bounds(m::AlpineNonlinearModel)

Resolve the bounds of the lifted variable using the information in lvartight and uvartight. This method only takes in known or trivial bounds information to reason lifted variable bound to avoid the cases of infinity bounds.

source

resolve_var_bounds(nonconvex_terms::Dict, discretization::Dict)

For discretization to be performed, we do not allow for a variable being discretized to have infinite bounds.
The lifted variables will have infinite bounds and the function infers bounds on these variables. This process
can help speed up the subsequent solve in subsequent iterations.

Only used in presolve bound tightening

source

Utility Methods

update_var_bounds
discretization_to_bounds
init_disc
to_discretization
flatten_discretization
add_adpative_partition
update_mip_time_limit
fetch_timeleft_symbol